+++ src/hkimjp/wil2/help.clj
@@ -7,20 +7,23 @@
   (page
    [:div
     [:div.text-2xl.font-medium "Help"]
-    [:div.mx-2
+    [:div.m-4
+     [:p.py-4 "今日の授業でなにやった？目を瞑ってじっと思い出してみ。"
+      "思い出せないような受講の態度は最悪。"
+      "記憶にもないことをコピペで取り繕っても最悪は変わりない。"
+      "わかったフリ、やったフリするのをやめよう。"]
      [:div [:span.font-bold "todays"]
-      [:ul.mx-4.list-disc
+      [:ul.mx-8.list-disc
        [:li "今日の WIL 未提出時"
-        [:p "upload ボタンを表示する"]]
+        [:p "upload ボタンを表示する。マークダウンファイルを選んで送信。"]]
        [:li "今日の WIL を提出済み"
-        [:p "WIL 提出済みのアカウントをリストする。
-             アカウントをクリックするとその人の WIL を表示するので
-             いいね、まあまあ、悪いねの点数をつける。"]]]]
+        [:p "WIL 提出済みのアカウントをリストする。"
+         "アカウントをクリックするとその人の WIL を表示する。"]
+        [:p "読んで点数をつける。(good:+2, soso:+1, bad:-1)"]]]]
      [:div [:span.font-bold "weeks"]
       [:p.mx-4 "授業の日付のリスト。クリックするとその日に届いた WIL を表示。"]]
      [:div [:span.font-bold "my"]
       [:p.mx-4 "自分の WIL 送信ポイント、受信ポイントを表示。"]]
-
      [:div [:span.font-bold "logout"]
       [:p.mx-4 "WIL をログアウト。"]]
      [:div [:span.font-bold "HELP"]